Added support for specifying a universal include folder.
To enable, add the following preference to your preferences.txt file:

preproc.includes_path=/path/to/include/folder

When this preference is specified, any .pde or .java files included
in the includes folder will be available when compiling sketch files
from the IDE.

This allows you to place commonly used code and files in a single folder
and have them avaiable to all of your Processing projects, without
having to copy the files into each project directory.

If the preference is not specified, is empty, or does not point to
a folder, it will be ignored.

Note, I basically reused code for parsing SketchCode file in Sketch.java.

Ideally, this code should be refactored in a way that it can be reused (so we don't
have two copies of the code).